telegram-flight-bot/requirements.txt
schedule==1.1.0
requests==2.26.0
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
selenium==4.15.2